feat: ignore proxy configuration locally if no valid token or password is found
We use the proxy configuration in all the templates, as we want users to run behind a proxy all the times.
This PR changes the internal mechanism to gracefully ignore the proxy configuration locally when we don't
see a valid token or proxy password. This way, new users can run their actors locally without being logged in.

Closes #262