fix: respect `forceCloud` in `KeyValueStore.getPublicUrl()` calls (#462)
The original check from #302 was not sufficient. When the user passes
`forceCloud: true` to `KVS.open()`, the SDK instance points to a
platform-backed KVS, even if the script is run locally.

This PR proposes checking the runtime type of `KeyValueStore.client`. If
this is a client instance from `apify-client`, we're communicating with
the platform, regardless of the place of execution (Platform / locally).

Closes #459